backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
torch==1.10.0
torchvision==0.11.1
numpy==1.21.4
gunicorn==20.1.0